Register the styled namespace
`lib/syntaxes/styled/` is the first syntax registered beside the core: every rule of the registry now also stands under `@stylistic/styled/<rule>`, built over an adapter that reads a styled template (`postcss-styled-syntax`) — its `accepts` takes a root carrying the parser's `styledSyntaxRangeStart` mark, and plain CSS besides, since the namespace is a superset of the core; the embedding of a template is answered here now, `lineAt` and its throw included, and the core borrows those answers until the branch that turns it away from styled templates. The namespace's README says what it is and shows the `overrides` a project writes for the files that carry templates.

The styled cases of `indentation` move with the reading, to `lib/syntaxes/styled/rules/indentation/index.test.ts`, and run under the namespace's own rule name; the registry's test checks every rule's name and messages under the namespace as it always meant to, now over a list that is no longer empty. The break-reading debt entry of `lineAt` follows the line to the adapter.

Run over the change: `make verify`; `make oracles` against `origin/main`, all six oracles with no row added, removed or changed — the oracles read css, scss and less, so the styled reading is answered by the namespace's moved cases and by nothing else.
