+++ scripts/check-break-readings.ts
@@ -76,7 +76,7 @@ const DEBT: Record<string, string[]> = {
 		`if (source.slice(index, index + 2) === \`\\r\\n\`) return`,
 	],
 	"lib/rules/indentation/index.ts": [`target: \`\\n\`,`],
-	"lib/syntaxes/css/index.ts": [`let found = text.split(\`\\n\`)[line - 1]`],
+	"lib/syntaxes/styled/index.ts": [`let found = text.split(\`\\n\`)[line - 1]`],
 	"lib/rules/max-empty-lines/index.ts": [`target: CRLF.test(rootString) ? \`\\r\\n\` : \`\\n\`,`],
 	"lib/rules/max-line-length/index.ts": [
 		`styleSearch({ source: rootString, target: [\`\\n\`], comments: \`check\` }, (match) => checkNewline(match))`,
